添加选择逻辑服务

+++ game-frame-server/game-network-message/src/main/java/com/xinyue/network/message/common/GameMessageMetaData.java
@@ -11,6 +11,6 @@ import com.xinyue.network.EnumServerType;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface GameMessageMetaData {
 	public EnumServerType serverType();
-	public int id();
+	public short id();
 	public GameMessageType type();
 }
